Auth tests now retrieves $token and $auth information
Split GeneralQueryTests
