+++ tests/Driver.Tests/Queries/GeneralQueryTests.cs
@@ -21,8 +21,7 @@ public abstract class GeneralQueryTests<T>
     public GeneralQueryTests(ITestOutputHelper logger) {
         Logger = logger;
     }
-
-
+    
     private record GroupedCountries {
         string? country;
         string? total;
@@ -36,243 +35,6 @@ public abstract class GeneralQueryTests<T>
     private class MathResultDocument {
         public float result {get; set;}
     }
-
-    [Fact]
-    public async Task StopStartConnectionTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            string sql = "INFO FOR DB;";
-            var response = await db.Query(sql, null);
-            Assert.NotNull(response);
-            TestHelper.AssertOk(response);
-
-            await db.Close();
-            await Assert.ThrowsAsync<InvalidOperationException>(async () => await db.Query(sql, null));
-            db.Dispose();
-            await Assert.ThrowsAsync<InvalidOperationException>(async () => await db.Query(sql, null));
-
-            db = new();
-            await db.Open(TestHelper.Default);
-            
-            response = await db.Query(sql, null);
-            Assert.NotNull(response);
-            TestHelper.AssertOk(response);
-        }
-    );
-
-    
-    [Fact]
-    public async Task SignInRootAuthTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            var signinObject = new RootAuth(TestHelper.User, TestHelper.Pass);
-            var response = await db.Signin(signinObject);
-            Assert.NotNull(response);
-            TestHelper.AssertOk(response);
-            Assert.True(response.TryGetResult(out Result result));
-            string? signinJwt = result.GetObject<string>();
-            signinJwt.Should().BeNullOrEmpty();
-        }
-    );
-
-    [Fact]
-    public async Task SignInNamespaceUserTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            var user = "DatabaseUser";
-            var password = "TestPassword";
-
-            string sql = $"DEFINE LOGIN {user} ON NAMESPACE PASSWORD '{password}';";
-            var queryResponse = await db.Query(sql, null);
-            Assert.NotNull(queryResponse);
-            TestHelper.AssertOk(queryResponse);
-
-            var signinObject = new NamespaceAuth( user, password, TestHelper.Namespace );
-            var signinResponse = await db.Signin(signinObject);
-            Assert.NotNull(signinResponse);
-            TestHelper.AssertOk(signinResponse);
-            Assert.True(signinResponse.TryGetResult(out Result result));
-            string? signinJwt = result.GetObject<string>();
-            signinJwt.Should().NotBeNullOrEmpty();
-        }
-    );
-
-    [Fact]
-    public async Task SignInDatabaseUserTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            var user = "DatabaseUser";
-            var password = "TestPassword";
-
-            string sql = $"DEFINE LOGIN {user} ON DATABASE PASSWORD '{password}';";
-            var queryResponse = await db.Query(sql, null);
-            Assert.NotNull(queryResponse);
-            TestHelper.AssertOk(queryResponse);
-
-            var signinObject = new DatabaseAuth( user, password, TestHelper.Namespace, TestHelper.Database );
-            var signinResponse = await db.Signin(signinObject);
-            Assert.NotNull(signinResponse);
-            TestHelper.AssertOk(signinResponse);
-            Assert.True(signinResponse.TryGetResult(out Result result));
-            string? signinJwt = result.GetObject<string>();
-            signinJwt.Should().NotBeNullOrEmpty();
-        }
-    );
-    
-    [Fact]
-    public async Task SignUpAndSignInScopedUserTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            var user = "TestUser@example.com";
-            var password = "TestPassword";
-            var scope = "account";
-
-            string sql = $"DEFINE SCOPE {scope}\n"
-              + "   SIGNIN ( SELECT * FROM user WHERE email = $user AND crypto::argon2::compare(password, $pass) )\n"
-              + "   SIGNUP ( CREATE user SET email = $user, password = crypto::argon2::generate($pass) )\n"
-              + ";";
-            var queryResponse = await db.Query(sql, null);
-            Assert.NotNull(queryResponse);
-            TestHelper.AssertOk(queryResponse);
-            
-            var signupObject = new ScopeAuth( user, password, TestHelper.Namespace, TestHelper.Database, scope );
-            var signupResponse = await db.Signup(signupObject);
-            Assert.NotNull(signupResponse);
-            TestHelper.AssertOk(signupResponse);
-            Assert.True(signupResponse.TryGetResult(out Result signupResult));
-            string? signupJwt = signupResult.GetObject<string>();
-            signupJwt.Should().NotBeNullOrEmpty();
-
-            var signinObject = new ScopeAuth( user, password, TestHelper.Namespace, TestHelper.Database, scope );
-            var signinResponse = await db.Signin(signinObject);
-            Assert.NotNull(signinResponse);
-            TestHelper.AssertOk(signinResponse);
-            Assert.True(signinResponse.TryGetResult(out Result result));
-            string? signinJwt = result.GetObject<string>();
-            signinJwt.Should().NotBeNullOrEmpty();
-        }
-    );
-    
-    [Fact]
-    public async Task SignUpAndSignInScopedUserWithDefinedIdTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            var user = "TestUser@example.com";
-            var password = "TestPassword";
-            var scope = "account";
-            var id = "user:123";
-
-            string sql = $"DEFINE SCOPE {scope}\n"
-              + "   SIGNIN ( SELECT * FROM user WHERE email = $user AND crypto::argon2::compare(password, $pass) )\n"
-              + "   SIGNUP ( CREATE $id SET email = $user, password = crypto::argon2::generate($pass) )\n"
-              + ";";
-            var queryResponse = await db.Query(sql, null);
-            Assert.NotNull(queryResponse);
-            TestHelper.AssertOk(queryResponse);
-            
-            var signupObject = new IdScopeAuth( id, user, password, TestHelper.Namespace, TestHelper.Database, scope );
-            var signupResponse = await db.Signup(signupObject);
-            Assert.NotNull(signupResponse);
-            TestHelper.AssertOk(signupResponse);
-            Assert.True(signupResponse.TryGetResult(out Result signupResult));
-            string? signupJwt = signupResult.GetObject<string>();
-            signupJwt.Should().NotBeNullOrEmpty();
-
-            var signinObject = new ScopeAuth( user, password, TestHelper.Namespace, TestHelper.Database, scope );
-            var signinResponse = await db.Signin(signinObject);
-            Assert.NotNull(signinResponse);
-            TestHelper.AssertOk(signinResponse);
-            Assert.True(signinResponse.TryGetResult(out Result result)); 
-            string? signinJwt = result.GetObject<string>();
-            signinJwt.Should().NotBeNullOrEmpty();
-        }
-    );
-
-    [Fact]
-    public async Task SwitchDatabaseTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            var nsName = db.GetConfig().Namespace!;
-            var originalDbName = db.GetConfig().Database!;
-            var otherDbName = "DifferentDb";
-
-            TestObject<int, string> expectedOriginalObject = new(1, originalDbName);
-            TestObject<int, string> expectedOtherObject = new(1, otherDbName);
-            
-            Thing thing = Thing.From("object", expectedOriginalObject.Key.ToString());
-            await db.Create(thing, expectedOriginalObject);
-
-            {
-                var useResponse = await db.Use(otherDbName, nsName);
-                Assert.NotNull(useResponse);
-                TestHelper.AssertOk(useResponse);
-
-                await db.Create(thing, expectedOtherObject);
-
-                var response = await db.Select(thing);
-
-                Assert.NotNull(response);
-                TestHelper.AssertOk(response);
-                Assert.True(response.TryGetResult(out Result result));
-                TestObject<int, string>? doc = result.GetObject<TestObject<int, string>>();
-                doc.Should().BeEquivalentTo(expectedOtherObject);
-            }
-
-            {
-                var useResponse = await db.Use(originalDbName, nsName);
-                Assert.NotNull(useResponse);
-                TestHelper.AssertOk(useResponse);
-                
-                var response = await db.Select(thing);
-
-                Assert.NotNull(response);
-                TestHelper.AssertOk(response);
-                Assert.True(response.TryGetResult(out Result result));
-                TestObject<int, string>? doc = result.GetObject<TestObject<int, string>>();
-                doc.Should().BeEquivalentTo(expectedOriginalObject);
-            }
-
-        }
-    );
-
-    [Fact]
-    public async Task SwitchNamespaceTest() => await DbHandle<T>.WithDatabase(
-        async db => {
-            var originalNsName = db.GetConfig().Namespace!;
-            var otherNsName = "DifferentNs";
-            var dbName = db.GetConfig().Database!;
-
-            TestObject<int, string> expectedOriginalObject = new(1, originalNsName);
-            TestObject<int, string> expectedOtherObject = new(1, otherNsName);
-            
-            Thing thing = Thing.From("object", expectedOriginalObject.Key.ToString());
-            await db.Create(thing, expectedOriginalObject);
-
-            {
-                var useResponse = await db.Use(dbName, otherNsName);
-                Assert.NotNull(useResponse);
-                TestHelper.AssertOk(useResponse);
-
-                await db.Create(thing, expectedOtherObject);
-
-                var response = await db.Select(thing);
-
-                Assert.NotNull(response);
-                TestHelper.AssertOk(response);
-                Assert.True(response.TryGetResult(out Result result));
-                TestObject<int, string>? doc = result.GetObject<TestObject<int, string>>();
-                doc.Should().BeEquivalentTo(expectedOtherObject);
-            }
-
-            {
-                var useResponse = await db.Use(dbName, originalNsName);
-                Assert.NotNull(useResponse);
-                TestHelper.AssertOk(useResponse);
-
-                var response = await db.Select(thing);
-
-                Assert.NotNull(response);
-                TestHelper.AssertOk(response);
-                Assert.True(response.TryGetResult(out Result result));
-                TestObject<int, string>? doc = result.GetObject<TestObject<int, string>>();
-                doc.Should().BeEquivalentTo(expectedOriginalObject);
-            }
-
-        }
-    );
 
     [Fact]
     public async Task SimpleFuturesQueryTest() => await DbHandle<T>.WithDatabase(
